fixed radius slider problem, it now updates as intended
